requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
blessed==1.21.0
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-autoslug==1.9.9
django-dirtyfields==1.9.7
django-picklefield==3.3
django_vite_plugin==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
newsapi-python==0.2.7
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==3.5.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
